Fix Spector.js on Firefox by not initializing polyfill WebXR classes unless we are using WebXR
